refactor: code clean up related store and split type
