+++ client/src/store/reducer/thread.reducer.ts
@@ -5,75 +5,11 @@ import {
   createAsyncAction,
 } from 'typesafe-actions'
 import { AxiosError } from 'axios'
-
-interface UserType {
-  id: number
-  email: string
-  name: string
-  profileImageUrl: string
-}
-
-interface FileType {
-  id: number
-  url: string
-  type: string
-  createdAt: string
-  updatedAt: string
-  profileImageUrl: string
-}
-
-interface ReactionType {
-  id: number
-  content: string
-  User: UserType
-}
-
-export interface ThreadType {
-  id: number
-  createdAt: string
-  updatedAt: string
-  messageCount: number
-  profileImageUrl: string[] | null
-  User: UserType
-}
-
-export interface MessageType {
-  id: number
-  content: string
-  isHead: true
-  createdAt: string
-  updatedAt: string
-  User: UserType
-  File: FileType[]
-  Reactions: ReactionType[]
-}
-
-export interface GetThreadResponseType {
-  id: number
-  createdAt: string
-  updatedAt: string
-  User: UserType
-  headMessage: MessageType
-  replyCount: number
-  userProfileMax5: string[]
-  commenterCount: number
-  lastReplyTime: string
-}
-
-export interface CreateThreadRequestType {
-  content: string
-  channelId: number
-  fileInfoList: { filePath: string; type: string }[] | null
-}
-
-export interface GetChannelInfoResponseType extends Object {
-  id: number
-  type: string
-  name: string
-  createdAt: string
-  updatedAt: string
-  user: UserType[]
-}
+import {
+  GetThreadResponseType,
+  GetChannelInfoResponseType,
+  CreateThreadRequestType,
+} from '@type/thread.type'
 
 interface ThreadState {
   channelInfo: GetChannelInfoResponseType
@@ -82,62 +18,6 @@ interface ThreadState {
   error: AxiosError | null
 }
 
-// Action
-export const GET_THREADS = 'thread/GET_THREADS' as const
-export const GET_THREADS_SUCCESS = 'thread/GET_THREADS_SUCCESS' as const
-export const GET_THREADS_ERROR = 'thread/GET_THREADS_ERROR' as const
-export const CREATE_THREAD = 'thread/CREATE_THREAD' as const
-export const RECEIVE_CREATE_THREAD = 'thread/RECEIVE_CREATE_THREAD' as const
-export const GET_CHANNEL_INFO = 'thread/GET_CHANNEL_INFO' as const
-export const GET_CHANNEL_INFO_SUCCESS = 'thread/GET_CHANNEL_INFO_SUCCESS' as const
-export const GET_CHANNEL_INFO_ERROR = 'thread/GET_CHANNEL_INFO_ERROR' as const
-
-// Action generator
-export const getThreads = createAction(GET_THREADS)<number>()
-export const getThreadsSuccess = createAction(GET_THREADS_SUCCESS)<
-  GetThreadResponseType[]
->()
-export const getThreadsError = createAction(GET_THREADS_ERROR)<AxiosError>()
-export const createThread = createAction(CREATE_THREAD)<
-  CreateThreadRequestType
->()
-export const receiveCreateThread = createAction(RECEIVE_CREATE_THREAD)<
-  GetThreadResponseType
->()
-export const getChannelInfo = createAction(GET_CHANNEL_INFO)<number>()
-export const getChannelInfoSuccess = createAction(GET_CHANNEL_INFO_SUCCESS)<
-  GetChannelInfoResponseType
->()
-export const getChannelInfoError = createAction(GET_CHANNEL_INFO_ERROR)<
-  AxiosError
->()
-
-export const getThreadsAsync = createAsyncAction(
-  GET_THREADS,
-  GET_THREADS_SUCCESS,
-  GET_THREADS_ERROR,
-)<number, GetThreadResponseType[], AxiosError>()
-
-export const getChannelInfoAsync = createAsyncAction(
-  GET_CHANNEL_INFO,
-  GET_CHANNEL_INFO_SUCCESS,
-  GET_CHANNEL_INFO_ERROR,
-)<number, GetChannelInfoResponseType, AxiosError>()
-
-// action
-const actions = {
-  getThreads,
-  getThreadsSuccess,
-  getThreadsError,
-  createThread,
-  receiveCreateThread,
-  getChannelInfo,
-  getChannelInfoSuccess,
-  getChannelInfoError,
-}
-export type ThreadAction = ActionType<typeof actions>
-
-// initial state
 const initialState: ThreadState = {
   channelInfo: {
     id: 0,
@@ -152,8 +32,57 @@ const initialState: ThreadState = {
   error: null,
 }
 
+export const GET_THREADS_REQUEST = 'thread/GET_THREADS_REQUEST' as const
+export const GET_THREADS_SUCCESS = 'thread/GET_THREADS_SUCCESS' as const
+export const GET_THREADS_ERROR = 'thread/GET_THREADS_ERROR' as const
+export const CREATE_THREAD = 'thread/CREATE_THREAD' as const
+export const RECEIVE_CREATE_THREAD = 'thread/RECEIVE_CREATE_THREAD' as const
+export const GET_CHANNEL_INFO_REQUEST = 'thread/GET_CHANNEL_INFO_REQUEST' as const
+export const GET_CHANNEL_INFO_SUCCESS = 'thread/GET_CHANNEL_INFO_SUCCESS' as const
+export const GET_CHANNEL_INFO_ERROR = 'thread/GET_CHANNEL_INFO_ERROR' as const
+
+const getThreadsRequest = createAction(GET_THREADS_REQUEST)<number>()
+const getThreadsSuccess = createAction(GET_THREADS_SUCCESS)<
+  GetThreadResponseType[]
+>()
+const getThreadsError = createAction(GET_THREADS_ERROR)<AxiosError>()
+export const createThread = createAction(CREATE_THREAD)<
+  CreateThreadRequestType
+>()
+export const receiveCreateThread = createAction(RECEIVE_CREATE_THREAD)<
+  GetThreadResponseType
+>()
+const getChannelInfoRequest = createAction(GET_CHANNEL_INFO_REQUEST)<number>()
+const getChannelInfoSuccess = createAction(GET_CHANNEL_INFO_SUCCESS)<
+  GetChannelInfoResponseType
+>()
+const getChannelInfoError = createAction(GET_CHANNEL_INFO_ERROR)<AxiosError>()
+export const getThreadsAsync = createAsyncAction(
+  GET_THREADS_REQUEST,
+  GET_THREADS_SUCCESS,
+  GET_THREADS_ERROR,
+)<number, GetThreadResponseType[], AxiosError>()
+export const getChannelInfoAsync = createAsyncAction(
+  GET_CHANNEL_INFO_REQUEST,
+  GET_CHANNEL_INFO_SUCCESS,
+  GET_CHANNEL_INFO_ERROR,
+)<number, GetChannelInfoResponseType, AxiosError>()
+
+const actions = {
+  getThreadsRequest,
+  getThreadsSuccess,
+  getThreadsError,
+  createThread,
+  receiveCreateThread,
+  getChannelInfoRequest,
+  getChannelInfoSuccess,
+  getChannelInfoError,
+}
+
+export type ThreadAction = ActionType<typeof actions>
+
 const reducer = createReducer<ThreadState, ThreadAction>(initialState, {
-  [GET_THREADS]: (state, action) => ({
+  [GET_THREADS_REQUEST]: (state, _) => ({
     ...state,
     loading: true,
     error: null,
@@ -174,7 +103,7 @@ const reducer = createReducer<ThreadState, ThreadAction>(initialState, {
     ...state,
     threadList: [...state.threadList, action.payload],
   }),
-  [GET_CHANNEL_INFO]: (state, action) => ({
+  [GET_CHANNEL_INFO_REQUEST]: (state, _) => ({
     ...state,
     loading: true,
     error: null,
