+++ client/src/type/user.type.ts
@@ -1,0 +1,6 @@
+export interface UserType {
+  id: number
+  email: string
+  name: string
+  profileImageUrl: string
+}
